IRMG Gunsprite update and IRMG marksman rifle (#1973)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request


![inteq_guns_2_nobg](https://github.com/shiptest-ss13/Shiptest/assets/60533805/83741b52-d15b-4cf3-a8ba-4426abd6cda3)

![ssg04](https://github.com/shiptest-ss13/Shiptest/assets/60533805/c1fe2b34-3598-409b-9a80-07e9f17fff8f)

Updates sprites for all IRMG ballistics to look prettier and less muddy.
Also introduces the SsG-04, an Inteq modification of the CM-GAL-S
marksman rifle. Doesn't change the loadout of any Inteq ships to include
the new rifle, that will come later as part of some planned remapping.

## Why It's Good For The Game

Sprite improvements are good and IRMG had a missing spot in their gun
roster.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added the Inteq SsG-04 marksman rifle
tweak: Improved IRMG gunsprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->