Overmap 4.5: Wasteplanet Expansion (#1945)
that is to say creates biomes for waste planets

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

![dreamseeker_3fbK5Yfp2k](https://user-images.githubusercontent.com/94164348/236656350-041bfbb6-fea0-42d1-b7e1-f75c4c1159ac.png)


![dreamseeker_KHidB6LdRo](https://user-images.githubusercontent.com/94164348/236656352-c0320b1f-b542-4525-a2c0-62f0e65645d8.png)


https://cdn.discordapp.com/attachments/881319909991260180/1104612113634308217/dreamseeker_Bx4rB7Eq45.mp4

Moves waste planets over to multibiome planet generation while keeping
most of their charm intact. It's not "complete" but that's because the
hivebot stuff is lacking (which I plan on addressing next), and didn't
really put much thought into swarmers.
Also I'm not a spriter
## Why It's Good For The Game
Standardizes waste planets with the rest of our planets while keeping
their character.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Waste planets now generate with biomes, please report any
weirdness.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->