Ports a tasty treat (the BLT) (#2005)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
As said above, this PR ports the Bisexual Lesbian Transgender Sandwich,
or most commonly known as "The BLT" from TG
its a tasty treat :)


![image](https://github.com/shiptest-ss13/Shiptest/assets/79304582/b320b21e-c0ac-4209-9088-372f726911fa)


![image](https://github.com/shiptest-ss13/Shiptest/assets/79304582/c90b128b-de32-42f3-8142-92581105951e)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
more food is always good!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Ports the BLT from TG
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->