Pixel shifting out of the way port from Skyrat (#2024)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Ports pixel shifting out of the way from skyrat.
https://github.com/Skyrat-SS13/Skyrat-tg/pull/15175
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Makes it way more convenient to players that are pixel shifted won't get
pushed around except from the direction they shifted in. This makes so
those players that usually talking near a wall don't block the way or
gets their RP interrupted.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Allows pixel shifted players to not be pushed around depending on
direction.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Niko <caio9009@hotmail.com>