The Tidal Barrier (#2015)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the HighTide and TrunkTide. The original flavor of Tide remains,
for now, until its replacement is done.

## Why It's Good For The Game

i am getting very, very tired of manifests with three or more hightides
on them. even with proper ship limits, these ships were all on the
cutting board in the first place due to their lackluster design and the
playstyles they tended to encourage.

## Changelog

:cl:
del: Removed the HighTide and TrunkTide
del: deprecated the Tide
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->