Adds an overlay and sound to cleaning (#2035)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds a bubble cleaning effect (from TG) as an overlay to stuff that is
being cleaned. Also adds a sound when starting to clean. Works on soap,
mops, and rags.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes it more obvious what are people are cleaning!

https://github.com/shiptest-ss13/Shiptest/assets/132854285/7ba81225-337a-46ff-ab9c-5e4ad5787854

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: Nitha
add: Cleaning now has an effect and sound.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->