last modern railing fix (#2040)
## About The Pull Request

adjusts the modern railing sprite for the god knows which time to
finally align perfectly at all situations

## Why It's Good For The Game

ugh

![image](https://github.com/shiptest-ss13/Shiptest/assets/105491762/88c15165-efe6-4abf-8e16-8c39a03d8322)

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: tweaked the modern railing sprite
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->