Beluga-Class Transport (we rich as hell!) (#1920)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
so, what is the Beluga class?
well! its a high-class transport ship, used by the rich and important.
its sleek design, rich interior and high speed makes it perfect for
transportation of high value passengers.

While the ship has no science materials, barely any medical supplies,
almost no security gear or engineering supplies, it DOES have a fancy
kitchen, a rich interior design and a total of... 6 cryo pods!!! (rich
people don't SHARE cryopods???)

it exists to fill the niche of "rich person roleplay", something that
was left empty after the removal of the high class

----

the ship contains:
two bedrooms
a bunk for the crew
a large meeting room
a kitchen/bar
a inteq security checkpoint
and other essentials!

----

The crew:
1x Captain
1x Bridge officer
2x Inteq private security officers
1x bartender
1x mechanic
1x janitor
4x high class passenger


![image](https://user-images.githubusercontent.com/79304582/233655271-d4ab311a-d302-4b7a-9021-9e1998621c0f.png)

![2023 04 21-15 59
16](https://user-images.githubusercontent.com/79304582/233655325-69945d73-003c-4cba-99c3-9840e6adfb43.png)


![image](https://user-images.githubusercontent.com/79304582/233712187-633ceb13-46a7-4bd7-87d2-79148fef1a9f.png)


![image](https://user-images.githubusercontent.com/79304582/233712222-422c249e-c716-404a-8d4d-c4f6306f7ff1.png)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
its a unique ship that fills a currently un-used niche, and it builds
onto the lore by having hired inteq guards
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Adds the Beluga class transport
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: spockye <79304582+spockye@users.noreply.github.com>
Co-authored-by: Bjarl <94164348+Bjarl@users.noreply.github.com>