Updates the Talos (#2032)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

![talos_2](https://github.com/shiptest-ss13/Shiptest/assets/60533805/ef236615-f91f-4a65-9874-0226d78b624f)

Updates the Talos. 

- added IRMG space suits
- replaced pump shotguns with Mastiffs and small magazines
- reduced burn chamber to 1x3 tiles
- reduced Recruit slots to 5
- increased Corpsman slots to 2

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
tweak: Updated Talos with new assets, adjusted engine and job slots
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->