Remaps the Colossus's armory (#2030)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request


![colossus_2](https://github.com/shiptest-ss13/Shiptest/assets/60533805/2c6f8ec6-cce2-4078-b2b3-140f861e44cf)

Completely redesigns the Colossus's starboard side, adds a
Master-at-Arms role to the Colossus, and adds an IRMG Comms Relay to the
bridge.

## Why It's Good For The Game

Colossus needed a little love and new IRMG assets need a chance to
shine.

## Changelog

:cl:
add: Added a Master-at-Arms to the Colossus
tweak: Redesigned the Colossus's armory
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->