Blots Out the Sun (Solar Removal) (#2078)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the Solar-class.

## Why It's Good For The Game

The Solar is bland, outdated, overpowered by virtue of having an
omnilathe, and sits in a niche that is already filled by better ships in
the indie catalogue. It has been slowly fermenting in the "we need to
rework this" category for nearly a year without anybody bothering to fix
it.

## Changelog

:cl:
del: Removed the Solar-class
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->