Ship tags and descriptions (#2010)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This adds descriptions and tags to every ship in the game that is not a
subshuttle. It also makes these viewable in the Ship Purchase, Ship
Join, and Shuttle Manipulator tabs. It also reworks the shuttle
manipulator template tab, and adds the ability to add or remove tags
from ships manually in the ship editor.

![6dd8e53c4b8224837f1d9748c4a37ccf](https://github.com/shiptest-ss13/Shiptest/assets/55075747/f1890609-6140-4a1e-bcb0-9827284ff676)

![77def6942b2f4c4fa4424f57b80f9058](https://github.com/shiptest-ss13/Shiptest/assets/55075747/20cfdbc3-2089-44fa-b4c0-d27511daa076)

![cc78d07d7c2e27526ab4c3756dcb7831](https://github.com/shiptest-ss13/Shiptest/assets/55075747/097a1a31-ba8e-44bd-9cdd-6a4565cc4eea)

![eab030d58d6187c8111162a141f761f5](https://github.com/shiptest-ss13/Shiptest/assets/55075747/fa76f828-a701-4799-9af6-17de36f08934)

Search features and other ship purchasing tgui related changes will be
added in separate PR.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This allows people to get a better sense of what they're buying. It also
reworks some years old shuttle manipulator ui.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:phoaly
add: Added descriptions to all jsons
add: Added tags to all ship jsons
add: Added Ship Tags and descriptions to the ship purchase/join/shuttle
manipulator menus
add: Added tag manipulation to the ship editor
tweak: Tweaks the UI of the shuttle manipulator
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Latency <55075747+Latentish@users.noreply.github.com>