Junker-Class Salvage(d) Ship (#2046)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds the Junker-class, a refinement on the older Tide-class maps, more
suited for an RP environment, while still executing on the aesthetic.
![2023-06-23 15 52
47](https://github.com/shiptest-ss13/Shiptest/assets/78622038/eeaa9906-7e19-4458-b40c-62953200f455)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Maptainers have long wanted to remove the tides, and replace them with a
suitable ship. I hope this can be that ship.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: added the Junker-class salvage(d) ship
add: added the reployer
del: removed the trunktide (older version of the junker)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->