Stabilizes space_level handling, preventing overlapping planet generation/allocation (#2097)
## About The Pull Request

During a recent uptime, the bottom-left corner of a z-level became
"cursed". New planets would all generate there, regardless of whether a
previous planet already had. When any of the planets unloaded
themselves, they would unload their reserve with them, deleting the
planet unexpectedly. The other planets which had also generated in the
same spot were left behind, appearing to be empty space: in truth, they
were the gap left behind after a virtual level is cleared.

As I wasn't able to diagnose the exact cause of the bug during the round
it occurred, this PR is admittedly a little bit of a shot in the dark.
My initial guess was that there was a null living in the list of
virtual_levels on that space_level. Although there weren't runtimes to
back this up, I've preemptively fixed that potential bug anyway. Another
possibility, which I have also addressed, was downstream of a race
condition resulting from the presence of a CHECK_TICK in add_new_zlevel:
I have eliminated the race condition by removing the sleep.

## Why It's Good For The Game

Planets overwriting other planets, and deleting people's ships, is bad.
That shouldn't happen.

## Changelog

:cl:
fix: Different planets should hopefully no longer repeatedly generate in
the exact same part of the map, overwriting each other and causing
double-dock SGTs.
/:cl: