Ensures external libraries can't be called by callbacks (#71346)
It is possible to create a callback whose `object` (the datum it tries
to call) is a path to an external library. Needless to say, it's
probably a bad idea to allow admins to call arbitrary external
libraries. Var-edited callbacks won't be able to reach the point where
the library is executed, but only because `WrapAdminProcCall` runtimes
when trying to call `CanProcCall` on a string, but if there is some way
to call a function that creates a non-varedited callback with an
external library as its object, this PR prevents that.

See above.

:cl:
admin: Admins are unable to invoke functions from external libraries
using callbacks.
/:cl: