chemistry "rebalance"  (#2060)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Moves the chem dispenser higher up the tech tree to coax people into
doing ghetto chem and to additionally help with that a new chemistry
cargo category is added to the outpost which contains full chemical jugs
of a few chemicals plus the starter kit crate from the medical category
and a tool crate. The category uses a new (soulful) chemical crate.

Also adds the previously mentioned chemical jugs which are 150u plastic
containers that cannot be stored in a backpack and have inconvenient
transfer amounts to "balance" how cheap they are. They also come with a
little fluff feature in the form of a tamper seal which lets you know
that no one touched anything inside the jug before you


![image](https://github.com/shiptest-ss13/Shiptest/assets/105491762/dc3f5db6-db50-4e28-b07d-e2d6da360fee)

![image](https://github.com/shiptest-ss13/Shiptest/assets/105491762/cc4675a1-317d-4067-bf18-7ba2aa80afe3)

![image](https://github.com/shiptest-ss13/Shiptest/assets/105491762/840e0307-4e21-4bf1-8c30-9b20c23f9848)

![image](https://github.com/shiptest-ss13/Shiptest/assets/105491762/4e2ef69f-f909-48f5-8acb-69eea4081b95)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Hopefully we'll see less chem dispensers around
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added chemical crates
add: Added chemical jugs
balance: moved the chem dispenser to applied bluespace research
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->