Autowiki Ships (#2096)
## About The Pull Request
Ports the rest of autowiki from tgstation/tgstation#64417, and adds our
own system for generating ship templates

## Why It's Good For The Game
Allows us to spend less time making redundant wiki pages for ships and
the like, gives players a rundown of ships without people having to
explicitly make pages

## Changelog

:cl: MothBlocks, MarkSuckerberg
add: The rest of the autowiki system, as well as the vending machine and
tech node
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->