Minor Lag-compensating Tweaks (#2050)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
A few minor tweaks to try and at least make the lag less damaging.

- Finishes porting tgstation/tgstation#59073 (since apparently we had
like half of it for some reason)
- Makes overmap movement timers their own timer subsystem
- Makes vox thermoregulator backpacks process on SSobj instead of
SSprocessing (it's higher priority)

## Why It's Good For The Game
The lag's bad and it's not easily fixed, so while we're fixing it we
might as well try to make it at least vaguely *playable* like this.

## Changelog

:cl:
tweak: Makes overmap movement and vox thermoregulators higher priority
processing targets
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->